requirements.txt
pandas 
openpyxl 
PyPDF2 
langchain 
chromadb==0.5.3 
groq 
python-dotenv 
requests 
sentence-transformers 
jsonpatch 
